backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyMuPDF>=1.23.0
google-generativeai>=0.3.0
python-multipart
numpy>=1.24.0
sqlalchemy>=2.0.0
passlib[bcrypt]
bcrypt>=4.0.0
pyjwt>=2.8.0
reportlab>=4.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pydantic[email]
python-jose[cryptography]>=3.3.0
starlette>=0.27.0